Drift mechanics
Tyre trail is left behind when turning and driving forwards at certain angles beyond threshold.
Only works while pressing the forwards key (W)
might change so that it works only when pressing shift, which is the handbrake. or make the hand brake apply momentary friction or impulse force in opposite direction to movement.
This game would be much more fun with a controller I think.